refactor(batch): extract shared sink/format writers into batch-sink.ts
Move the AST→file logic (formatOptionsFrom, docToExtracted, buildCombined,
downloadItem/Data, per-item + data.json + combined writers) out of batch.ts
into a folder-parameterized batch-sink.ts, so Fast Batch (ADR 0003) reuses
the exact same writers — only acquisition differs. batch.ts keeps the job
state machine and calls the shared writers with job.folder. No behavior
change to Standard batch; 180 tests green.
